refactor(memory): own generations in sync lifecycle
